Refactoring patches structure
