+++ README.md
@@ -13,12 +13,87 @@ gem 'patch_adams'
 
 And then execute:
 
-    $ bundle install
+```bash
+$ bundle install
+```
 
 Or install it yourself as:
 
-    $ gem install patch_adams
+```bash
+$ gem install patch_adams
+```
 
+## Loading
+
+### Core patches
+
+If you want to load the whole core patches, require `patch_adams` this way:
+
+```ruby
+require 'patch_adams/core'
+```
+
+If you want to cherry-pick specific core patches instead, require `patch_adams/minimal` and only those you need:
+
+```ruby
+require `patch_adams/minimal`
+require 'patch_adams/core/hash/diff'
+```
+
+### Extra patches
+
+Some patches require additional dependencies and are meant for specific contexts, and must be loaded independently.
+
+#### ActiveModel
+
+> :warning: You need to add `active_model >= 4.1` to your dependencies first
+
+If you want to require the whole `active_model` patches:
+
+```ruby
+require 'patch_adams/active_model'
+```
+
+If you want to cherry-pick specific `active_model` patches instead, require `patch_adams/minimal` and only those you need:
+
+```ruby
+require `patch_adams/minimal`
+require 'patch_adams/active_model/naming/name_model_as'
+```
+
+#### Rspec
+
+> :warning: You need to add `rspec >= 2` to your dependencies first
+
+If you want to require the whole `rspec` patches:
+
+```ruby
+require 'patch_adams/rspec'
+```
+
+If you want to cherry-pick specific `rspec` patches instead, require `patch_adams/minimal` and only those you need:
+
+```ruby
+require `patch_adams/minimal`
+require 'patch_adams/rspec/matches/json_eq'
+```
+
+#### Rack
+
+> :warning: You need to add `rack >= 1.5` to your dependencies first
+
+If you want to require the whole `rack` patches:
+
+```ruby
+require 'patch_adams/rack'
+```
+
+If you want to cherry-pick specific `rack` patches instead, require `patch_adams/minimal` and only those you need:
+
+```ruby
+require `patch_adams/minimal`
+require 'patch_adams/rack/request/subdominated'
+```
 
 ## Development
 
@@ -28,8 +103,16 @@ To install this gem onto your local machine, run `bundle exec rake install`. To 
 
 ## Contributing
 
-Bug reports and pull requests are welcome on GitHub at https://github.com/[USERNAME]/patch_adams. This project is intended to be a safe, welcoming space for collaboration, and contributors are expected to adhere to the [Contributor Covenant](http://contributor-covenant.org) code of conduct.
+Bug reports and pull requests are welcome on GitHub at https://github.com/mumuki/patch_adams. This project is intended to be a safe, welcoming space for collaboration, and contributors are expected to adhere to the [Contributor Covenant](http://contributor-covenant.org) code of conduct.
 
+## Adding new patches
+
+Patches are declared in separate files, each one containing a single patch or a group of very-related patches. This is done in order to make them easy to mantain, find and include in your project selectively.
+
+* If you want to add a core patch that does not require any additional dependency, place it in `lib/core` directory.
+* If you want to add an extra patch that requires an additional dependency, place it top-level in `lib` directory
+
+Additionally, if your patch is specific to a class or module - most are - place you patch file within a directory named as the class or module. Otherwise - that is, the patch is polymorphic and applies to many classes - place it top-level.
 
 ## License
