+++ lib/patch_adams/active_model.rb
@@ -1,0 +1,4 @@
+require_relative './minimal'
+require_relative './active_model/naming/demodulize_model_name'
+require_relative './active_model/naming/name_model_as'
+
