+++ lib/patch_adams/active_model/naming/demodulize_model_name.rb
@@ -1,12 +1,6 @@
 require 'active_model/naming'
 
 module ActiveModel::Naming
-  def name_model_as(other)
-    define_singleton_method :model_name do
-      other.model_name
-    end
-  end
-
   def demodulized_model_name
     define_singleton_method :model_name do
       ActiveModel::Name.new(self, nil, name.demodulize)
