+++ lib/patch_adams/core/array/merge_numbers.rb
@@ -1,0 +1,7 @@
+class Array
+  def merge_numbers!(key=:number)
+    self.each_with_index do |r, i|
+      r.send "#{key}=", i+1
+    end
+  end
+end
