+++ lib/patch_adams/core/array/to_csv.rb
@@ -1,0 +1,9 @@
+require 'csv'
+
+class Array
+  def to_csv
+    CSV.generate do |csv|
+      each { |hash| csv << hash.values }
+    end
+  end
+end
