+++ lib/patch_adams/core/array/to_h.rb
@@ -1,0 +1,5 @@
+class Array
+  def to_h
+    Hash[self]
+  end
+end
