+++ lib/patch_adams/core/as_module_name.rb
