+++ lib/patch_adams/core/as_variable_name.rb
