+++ lib/patch_adams/core/captures.rb
@@ -1,0 +1,9 @@
+class String
+  alias_method :captures, :gsub
+end
+
+class NilClass
+  def captures(*)
+    []
+  end
+end
