+++ lib/patch_adams/core/configurable.rb
