+++ lib/patch_adams/core/hash/ammend.rb
@@ -1,0 +1,5 @@
+class Hash
+  def amend(other)
+    other.merge(self)
+  end
+end
