+++ lib/patch_adams/core/hash/dig.rb
@@ -1,0 +1,12 @@
+class Hash
+  unless instance_methods.include?(:dig)
+    def dig(key, *keys)
+      value = self[key]
+      if !value.nil? && !keys.empty?
+        value.dig(*keys)
+      else
+        value
+      end
+    end
+  end
+end
