+++ lib/patch_adams/core/i18n/load_translations_path.rb
