+++ lib/patch_adams/core/json/pretty_parse.rb
