+++ lib/patch_adams/core/like.rb
