+++ lib/patch_adams/core/module/required.rb
@@ -1,0 +1,8 @@
+class Module
+  def required(name, message=nil)
+    message ||= "You need to implement method #{name}"
+    define_method name do |*|
+      raise message
+    end
+  end
+end
