+++ lib/patch_adams/core/object/defaulting.rb
@@ -1,0 +1,10 @@
+class Object
+  def defaulting(value)
+    if block_given? && !nil?
+      result = yield self
+    else
+      result = self
+    end
+    result || value
+  end
+end
