+++ lib/patch_adams/core/object/ensure_present.rb
@@ -1,0 +1,5 @@
+class Object
+  def ensure_present!(*args)
+    raise 'arguments must be non null' if args.any?(&:blank?)
+  end
+end
