+++ lib/patch_adams/core/object/to_stringified_h.rb
@@ -1,0 +1,5 @@
+class Object
+  def to_stringified_h
+    to_h.stringify_keys
+  end
+end
