+++ lib/patch_adams/core/regexp/matches.rb
