+++ lib/patch_adams/core/status.rb
