+++ lib/patch_adams/core/to_struct.rb
@@ -1,0 +1,12 @@
+class Hash
+  def to_struct
+    OpenStruct.new self
+  end
+end
+
+class Object
+  def struct(hash = {})
+    hash.to_struct
+  end
+end
+
