+++ lib/patch_adams/core/yaml/load_interpolated.rb
