+++ lib/patch_adams/minimal.rb
@@ -1,0 +1,5 @@
+module PatchAdams
+end
+
+require 'patch_adams/version'
+require 'active_support/all'
