--- lib/patch_adams/nil.rb
@@ -1,5 +1,0 @@
-class NilClass
-  def captures(*)
-    []
-  end
-end
