+++ lib/patch_adams/rack.rb
@@ -1,0 +1,2 @@
+require_relative './minimal'
+require_relative './rack/request/subdominated'
