+++ lib/patch_adams/rack/request/subdominated.rb
