--- lib/patch_adams/string.rb
@@ -1,3 +1,0 @@
-class String
-  alias_method :captures, :gsub
-end
