+++ spec/spec_helper.rb
@@ -2,6 +2,6 @@ require 'codeclimate-test-reporter'
 SimpleCov.start
 
 require 'patch_adams/rspec'
-require 'patch_adams'
-require 'patch_adams/request'
-require 'patch_adams/activemodel'
+require 'patch_adams/core'
+require 'patch_adams/rack'
+require 'patch_adams/active_model'
